VampPlugin wrapper hierarchy
